+++ caikit/core/data_model/prediction_status.py
